+++ monorepo/packages/wps-button-services-template/src/components/WPSButton.tsx
@@ -1,0 +1,23 @@
+import { ReactElement } from "react";
+import { Button } from "react-bootstrap";
+import { useGenericProviderContext } from '../GenericContextProvider';
+
+type Props = {
+    readonly text: string;
+};
+
+const WPSButton = ({ text }: Props): ReactElement => {
+    const newMap = useGenericProviderContext();
+    return (
+        <Button
+            onClick={
+                async () => { await newMap.merchant.getSearchResult().then((res) => 
+                    { res.data?.items?.length > 0 && window.open(res.data.items[0]?.link) }) }
+            }
+        >
+            {text}
+        </Button>
+    );
+};
+
+export default WPSButton;
