+++ monorepo/packages/wps-button-services-template/src/components/index.tsx
@@ -1,53 +1,48 @@
 import { FC } from 'react';
 import styled from 'styled-components';
 import Logo from 'assets/logo.svg';
-import { GenericContext, GenericProvider, useGenericProviderContext } from '../GenericContextProvider';
-import React, { ReactElement } from "react";
-import { Button } from "react-bootstrap";
+import { GenericProvider } from '../GenericContextProvider';
 import axios from 'axios';
-import types from "types";
+import lernaDemoTypes from "types";
+import WPSButton from './WPSButton';
 
-const userSvcObject: types.UserService = {
-    getUser: (id: string) : types.User => {
+const userSvcObject: lernaDemoTypes.UserService = {
+    getUser: (id: string) : lernaDemoTypes.User => {
         return {
-            id : "1",
+            id : id,
             name: "asdf"
         }
     },
-    findUser: (id: string) : types.User => {
+    findUser: (id: string) : lernaDemoTypes.User => {
         return {
-            id : "1",
+            id : id,
             name: "asdf"
         }
     }
 }
 
-const merchantSvcObject: types.MerchantService = {
-    getMerchant: (id: string) : types.Merchant => {
+const merchantSvcObject: lernaDemoTypes.MerchantService = {
+    getMerchant: (id: string) : lernaDemoTypes.Merchant => {
         return {
-            id : "1",
+            id : id,
             merchantname: "asdf"
         }
     },
-    findMerchant: (id: string) : types.Merchant => {
+    findMerchant: (id: string) : lernaDemoTypes.Merchant => {
         return {
-            id : "1",
+            id : id,
             merchantname: "asdf"
         }
     },
-    getSearchResult: async () : Promise<types.SearchObject> => {
-        console.log("hi!")
+    getSearchResult: async () : Promise<lernaDemoTypes.SearchObject> => {
         const callString = "https://www.googleapis.com/customsearch/v1?key=AIzaSyCqWcIPnxmD1yawBc2eNhVVgVs_uSI1Hyk&cx=8dedd47a8c8b92bb1&q=wish parcel service"
         const ret = {}
         try {
           const response = await axios.get(callString);
-          const resp2 = response as types.SearchObject;
-          console.log(resp2.data.items[0].link);
-          console.log(response);
-          return(response as types.SearchObject)
+          return(response as lernaDemoTypes.SearchObject)
         } catch (error) {
           console.error(error);
-          return(ret as types.SearchObject)
+          return(ret as lernaDemoTypes.SearchObject)
         }
     }
 }
@@ -57,29 +52,8 @@ const svcObjectMap = {
   merchant: merchantSvcObject
 };
 
-type Props = {
-  readonly text: string;
-};
-
-const WPSButton = ({ text }: Props): ReactElement => {
-  return (
-    <Button
-      onClick={() => {
-        window.open("http://parcel.wish.com");
-      }}
-    >
-      {text}
-    </Button>
-  );
-};
-
 /** Home page (demo) */
 const Home: FC = () => {
-  const context = React.useContext(GenericContext);
- // const context = (): typeof svcObjectMap => React.useContext(GenericContext);
-  // const {
-  //   svcObjectMap
-  // } = useGenericProviderContext();
   return (
     <GenericProvider m={svcObjectMap}>
       <StyledContainer>
@@ -89,10 +63,7 @@ const Home: FC = () => {
         <h3>Read the README for usage details! </h3>
         <img src={Logo} alt="React logo" />
       </StyledContainer>
-      <button onClick={context.merchant.getSearchResult}>
-        <h3>hello world</h3>
-      </button>
-      {/* <WPSButton text="hello world"></WPSButton> */}
+      <WPSButton text="hello world"></WPSButton>
     </GenericProvider>
   );
 };
