+++ monorepo/packages/wps-button-services-template/tsconfig.json
@@ -19,7 +19,7 @@
     ],
     "skipLibCheck": true,
     "allowSyntheticDefaultImports": true,
-    "strict": true,
+    "strict": false,
     "forceConsistentCasingInFileNames": true,
     "noFallthroughCasesInSwitch": true,
     "isolatedModules": true,
